+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.19.0
 * `NEW` VSCode: new setting `Lua.misc.parameters`
 * `NEW` quick fix: disable diagnostic in line/file
+* `CHG` intelli-scense: search from generic param to return
 * `CHG` improve performance
 * `CHG` text-document-synchronization: refactored
 * `FIX` missed syntax error `function m['x']() end`